+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+## v1.8.3 (2026-05-19)
+
+- Rewrote `Float2`, `Float3`, `Float4`, `Quaternion`, `Bool3`, `Trans` and `ColorRGBA`. Now they can be invoked without `new` keyword. `new` keyword is still recommended though.
+
 ## v1.8.2 (2026-04-28)
 
 - Added a new file `global.d.ts` and moved `IsUnique`, `ExpectTrue`, `$` into it.